backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-extensions==3.2.3
python-decouple==3.8
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.25.2
django-guardian==2.4.0
django-rest-auth==0.9.5
django-allauth==0.57.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-timezone-field==5.0
django-admin-interface==0.26.0
